requirements.txt
serpapi
python-dotenv
requests
beautifulsoup4
